Clarifies Cloud project time limit note
The word default in this note is wrong. It's not a default value that can be changed by a user, it's set. I am not sure if we want users to move to another tier but this is a first step to avoid the "okay, this is a default value, how do I change it" cycle.