Add GQL Customization to connect DB form [GDW-333]
## Description 🔖

This adds the ability to edit **GraphQL Field Customization** when either connecting or editing a db connection for an agent.

This was implemented using the newest form and form components:

<img width="548" alt="Screen Shot 2022-09-29 at 3 42 16 PM" src="https://user-images.githubusercontent.com/49927862/193155139-c0eb71fb-4829-42d9-8ec1-90582fe8054a.png">

## Solution and Design 🎨

This is a fairly minimal addition that makes use of the existing `react-hook-form` compatible `Form` and components. I had to edit the zod schema with the correct properties from the spec for GQL Sources.

I created a new dir for the `CustomizationForm` and implemented it in both the Connect and Edit components.

A quick note on a directory convention I am using here:

It's fairly common to have a components folder, and then within that, have sub-directories for more complicated components.

In that model, I have found it useful to use a `parts` directory as a next level of granularity within a component directory.
This helps to avoid cluttered components folder, by optionally adding a `parts` folder inside a component's folder for pieces that you want to break out but are NOT shared with anything else.

I've done this here with a fairly simple component `src/features/ConnectDB/components/Customization/parts/SectionHeader.tsx`, but one that benefits from having it's own component.

## Review Setup 💻

Video of feature:

https://www.loom.com/share/d737afca4be04da28f6943c1c8164d6f

### Tests

Added new customization form to storybook here ->> http://localhost:6006/?path=/story/data-connect-graphql-field-customization--primary

### Note on commit history

We had some issues trying to fix my accidental commital of a package-lock.json file and ended up doing some song and dance which involved @solomon-b ending up with credit for the first few commits. Just as a note for history, his name is on them, but it's my code, so if there's something to comment on, direct it to me.

PR-URL: https://github.com/hasura/graphql-engine-mono/pull/6134
Co-authored-by: Solomon <24038+solomon-b@users.noreply.github.com>
GitOrigin-RevId: 0c8891e246e5d048e70b6a8a91331f21049d1c8c